# public/openapi.yaml
openapi: 3.1.0
info:
  title: TASCO Atlas Hackathon API
  version: 0.1.0
  summary: Conversational discovery and deterministic map facade for the TASCO Atlas prototype.
  description: |
    REST contract for the camera-free TASCO Atlas hackathon prototype.

    All POI, profile, conversation, and evaluation data in this repository is synthetic.
    Search, ranking, and journey commerce fields are deterministic. OpenAI may optionally rewrite grounded chat prose. Route geometry and
    ETA values are simulated from supplied coordinates; they are not road- or traffic-aware.
servers:
  - url: http://localhost:3000
    description: Local Next.js development server
...
security: []
paths:
...
  /v1/search:
    get:
      tags: [Discovery]
      operationId: searchPlaces
      summary: Search and rank POIs using natural-language text.
      parameters:
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Language"
        - $ref: "#/components/parameters/ProfileId"
        - $ref: "#/components/parameters/LatitudeOptional"
        - $ref: "#/components/parameters/LongitudeOptional"
        - name: category
          in: query
          description: Exact or compatible POI category. When supplied, it is a hard filter.
          schema:
            type: string
          example: Quán cà phê
      responses:
        "200":
          description: Stable, deterministic ranking over the synthetic POI dataset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/chat:
    post:
      tags: [Conversation]
      operationId: chat
      summary: Turn a Vietnamese conversation turn into recommendations and a map action.
      description: |
        Ranking, journey actions, prices, totals, and revisions are deterministic and dataset-grounded. It has no durable session store.
        OpenAI may rewrite only assistant prose; provider failure uses the deterministic response.
        To maintain multi-turn context, pass the returned `sessionContext` (and optionally
        `history`) with the next request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              personalizedDate:
                summary: Personalized recommendation
                value:
                  sessionId: demo-001
                  profileId: U005
                  message: Gợi ý nơi hẹn hò lãng mạn tối nay ở Quận 1.
              clarification:
                summary: Ambiguous named place
                value:
                  sessionId: demo-002
                  message: Đưa tôi đến Galaxy.
      responses:
        "200":
          description: Natural-language response plus inspectable structured behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Missing message or invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    Query:
      name: q
      in: query
      required: true
      description: Natural-language place, address, category, or attribute query.
      schema:
        type: string
        minLength: 1
      example: cà phê yên tĩnh ở Quận 1
    Limit:
      name: limit
      in: query
      description: Maximum number of records. Values are clamped to 1–50.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
    Language:
      name: lang
      in: query
      description: Response/search locale metadata.
      schema:
        type: string
        default: vi
      example: vi
    ProfileId:
      name: profileId
      in: query
      description: Optional synthetic profile ID (`U001`–`U008`) used for ranking.
      schema:
        type: string
        pattern: "^U[0-9]{3}$"
      example: U001
...
    LatitudeOptional:
      name: lat
      in: query
      required: false
      description: Optional focus/origin latitude; used only when paired with `lon`.
      schema:
        type: number
        minimum: -90
        maximum: 90
    LongitudeOptional:
      name: lon
      in: query
      required: false
      description: Optional focus/origin longitude; used only when paired with `lat`.
      schema:
        type: number
        minimum: -180
        maximum: 180
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Coordinates:
      type: object
      additionalProperties: false
      required: [lat, lon]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lon:
          type: number
          minimum: -180
          maximum: 180
    Poi:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - category
        - brand
        - city
        - district
        - address
        - coordinates
        - rating
        - reviewCount
        - popularityScore
        - attributes
        - tags
        - description
        - datasetTier
      properties:
        id:
          type: string
          example: POI001
        name:
          type: string
          example: The Workshop Coffee
        category:
          type: string
          example: Quán cà phê
        brand:
          type: string
          example: The Workshop
        city:
          type: string
          example: TP.HCM
        district:
          type: string
          example: Quận 1
        address:
          type: string
          example: 27 Ngô Đức Kế, Bến Nghé, Quận 1
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        rating:
          type: number
          minimum: 0
          maximum: 5
        reviewCount:
          type: integer
          minimum: 0
        popularityScore:
          type: number
          minimum: 0
          maximum: 100
        attributes:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        datasetTier:
          type: string
          description: Synthetic fixture tier, for example `featured` or `generated`.
    PlaceResult:
      type: object
      additionalProperties: false
      required: [id, type, name, label, address, category, coordinates, source]
      properties:
        id:
          type: string
        type:
          type: string
          const: poi
        name:
          type: string
        label:
          type: string
        address:
          type: string
        category:
          type: string
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        distanceMeters:
          type: integer
          minimum: 0
        score:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
          const: tasco-dataset
        tags:
          type: array
          items:
            type: string
    SearchResponse:
      type: object
      additionalProperties: false
      required: [query, results, meta]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/PlaceResult"
        meta:
          type: object
          additionalProperties: false
          required: [limit, lang]
          properties:
            limit:
              type: integer
            lang:
              type: string
...
    ScoreBreakdown:
      type: object
      description: Weighted contributions to the final recommendation score.
      additionalProperties:
        type: number
    Recommendation:
      type: object
      additionalProperties: false
      required: [poi, score, reason]
      properties:
        poi:
          $ref: "#/components/schemas/Poi"
        score:
          type: number
          minimum: 0
          maximum: 1
        reason:
          type: string
        scoreBreakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        rewardPoints:
          type: integer
          description: Synthetic demo reward signal; no monetary value.
    ChatMessage:
      type: object
      additionalProperties: false
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    PendingClarification:
      type: object
      additionalProperties: false
      required: [entity, candidateIds]
      properties:
        entity:
          type: string
        candidateIds:
          type: array
          items:
            type: string
    SessionContext:
      type: object
      additionalProperties: false
      properties:
        sessionId:
          type: string
        profileId:
          type: string
        lastIntent:
          type: string
        lastQuery:
          type: string
        constraints:
          type: array
          items:
            type: string
        pendingClarification:
          $ref: "#/components/schemas/PendingClarification"
        journey:
          $ref: "#/components/schemas/JourneySessionState"
    JourneySessionState:
      type: object
      additionalProperties: false
      required: [query, location, selectedPoiIds, actionKinds, totalVnd, revision]
      properties:
        query: { type: string }
        location: { type: string }
        selectedPoiIds: { type: array, items: { type: string }, minItems: 2, maxItems: 3 }
        actionKinds: { type: array, items: { type: string, enum: [fuel, dining, parking] }, minItems: 2, maxItems: 3 }
        totalVnd: { type: integer, minimum: 0 }
        revision: { type: integer, minimum: 0 }
    ChatRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        profileId:
          type: string
          description: Optional synthetic user profile (`U001`–`U008`).
        location:
          $ref: "#/components/schemas/Coordinates"
        history:
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/ChatMessage"
        sessionId:
          type: string
        sessionContext:
          $ref: "#/components/schemas/SessionContext"
    MapAction:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          enum: [none, search, show, clarify, route, plan]
        query:
          type: string
        center:
          $ref: "#/components/schemas/Coordinates"
        zoom:
          type: number
        poiIds:
          type: array
          items:
            type: string
        selectedPoiId:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/PlaceResult"
        route:
          $ref: "#/components/schemas/RouteResult"
    PrivacyStatus:
      type: object
      additionalProperties: false
      required: [mode, persisted]
      properties:
        mode:
          type: string
          const: session-only
        persisted:
          type: boolean
          const: false
    ChatResponse:
      type: object
      additionalProperties: false
      required:
        - intent
        - assistantResponse
        - recommendations
        - confidence
        - mapAction
        - sessionContext
        - privacy
      properties:
        intent:
          type: string
          description: Detected intent or `clarification_required`.
        assistantResponse:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence in the selected response behavior, including a decision to clarify.
        mapAction:
          $ref: "#/components/schemas/MapAction"
        sessionContext:
          $ref: "#/components/schemas/SessionContext"
        privacy:
          $ref: "#/components/schemas/PrivacyStatus"
        journey:
          $ref: "#/components/schemas/Journey"
    JourneyAction:
      type: object
      additionalProperties: false
      required: [id, poiId, kind, miniApp, cta, reason, originalPriceVnd, discountVnd, finalPriceVnd, rewardPoints, status, simulated]
      properties:
        id: { type: string }
        poiId: { type: string }
        kind: { type: string, enum: [fuel, dining, parking] }
        miniApp: { type: string }
        cta: { type: string }
        reason: { type: string }
        originalPriceVnd: { type: integer, minimum: 0 }
        discountVnd: { type: integer, minimum: 0 }
        finalPriceVnd: { type: integer, minimum: 0 }
        rewardPoints: { type: integer, minimum: 0 }
        status: { type: string, enum: [ready, confirmed] }
        simulated: { type: boolean, const: true }
        sponsored:
          type: object
          additionalProperties: false
          required: [label, disclosure]
          properties:
            label: { type: string }
            disclosure: { type: string }
    Journey:
      type: object
      additionalProperties: false
      required: [id, title, actions, originalTotalVnd, discountTotalVnd, totalVnd, savingsVnd, rewardPoints, walletLabel, simulated, revision]
      properties:
        id: { type: string }
        title: { type: string }
        actions: { type: array, minItems: 2, maxItems: 3, items: { $ref: "#/components/schemas/JourneyAction" } }
        originalTotalVnd: { type: integer, minimum: 0 }
        discountTotalVnd: { type: integer, minimum: 0 }
        totalVnd: { type: integer, minimum: 0 }
        savingsVnd: { type: integer, minimum: 0 }
        rewardPoints: { type: integer, minimum: 0 }
        walletLabel: { type: string, const: Ví VETC }
        simulated: { type: boolean, const: true }
        revision:
          type: object
          additionalProperties: false
          required: [number, outcome, changedActionIds, message]
          properties:
            number: { type: integer, minimum: 0 }
            outcome: { type: string, enum: [composed, cheaper, no_cheaper_option] }
            changedActionIds: { type: array, items: { type: string } }
            message: { type: string }
    CoordinatePair:
      type: array
      description: GeoJSON coordinate tuple in `[longitude, latitude]` order.
      minItems: 2
      maxItems: 2
      items:
        type: number
    RouteGeometry:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: LineString
        coordinates:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/CoordinatePair"
    RouteManeuver:
      type: object
      additionalProperties: false
      required:
        - instruction
        - distanceMeters
        - durationSeconds
        - beginShapeIndex
        - endShapeIndex
        - streetNames
      properties:
        instruction:
          type: string
        distanceMeters:
          type: integer
          minimum: 0
        durationSeconds:
          type: integer
          minimum: 1
        beginShapeIndex:
          type: integer
          minimum: 0
        endShapeIndex:
          type: integer
          minimum: 0
        streetNames:
          type: array
          items:
            type: string
    RouteResult:
      type: object
      additionalProperties: false
      required: [routeId, sourceIndex, summary, geometry, maneuvers]
      properties:
        routeId:
          type: string
        sourceIndex:
          type: integer
          minimum: 0
        summary:
          type: object
          additionalProperties: false
          required: [distanceMeters, durationSeconds]
          properties:
            distanceMeters:
              type: integer
              minimum: 0
            durationSeconds:
              type: integer
              minimum: 1
        geometry:
          $ref: "#/components/schemas/RouteGeometry"
        maneuvers:
          type: array
          items:
            $ref: "#/components/schemas/RouteManeuver"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string